Further improve pretty formatting
XML formatting is now more conservative to account for whitespace having
meaning inside elements. For HTML, we format elements differently based
on whether they are flagged as "inline" or not. For example, a "strong"
element followed by text is now formatted like this:

    <strong>foo</strong>bar

Whereas before, it would be formatted like this:

    <strong>foo</strong>
    bar

The difference here is that in the second example, browsers add a space
before "bar", which results in the output looking different from what is
intended.

Changelog: fixed